backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-speech
google-cloud-texttospeech
google-cloud-aiplatform
